bugfix migrate

+++ cnme-api/database/migrations/2019_02_12_024449_drop_column_tipo_tarefa.php
@@ -14,8 +14,8 @@ class DropColumnTipoTarefa extends Migration
     public function up()
     {
         Schema::table('tarefas', function($table){
-          
-            $table->dropColumn('tipo');
+            if(Schema::hasColumn('tarefas', 'tipo'))
+                $table->dropColumn('tipo');
     
         });
     }
